pyproject.toml
[tool.poetry]
name = "lambda-sns"
version = "0.1.0"
description = ""
authors = ["amitkumarnag <amit.nag89@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "2.93.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
